+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - [4588](https://github.com/vegaprotocol/vega/pull/4588) - Fail when missing tendermint home and public key in `nodewallet import` command
 - [4623](https://github.com/vegaprotocol/vega/pull/4623) - Bug fix for `--snapshot.db-path` parameter not being used if it is set
 - [4634](https://github.com/vegaprotocol/vega/pull/4634) - Bug fix for `--snapshot.max-retries` parameter not working correctly
+- [4651](https://github.com/vegaprotocol/vega/pull/4651) - An array of fixes in the snapshot code path
 
 ## 0.47.5
 *2022-01-20*